+++ backend/src/main/java/com/bounswe2024group10/Tradeverse/dto/authentication/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.bounswe2024group10.Tradeverse.dto;
+package com.bounswe2024group10.Tradeverse.dto.authentication;
 
 public class LoginResponse {
     private boolean isSuccessful;
