+++ backend/src/main/java/com/bounswe2024group10/Tradeverse/dto/user/GetUserDetailsResponse.java
@@ -1,4 +1,4 @@
-package com.bounswe2024group10.Tradeverse.dto;
+package com.bounswe2024group10.Tradeverse.dto.user;
 
 public class GetUserDetailsResponse {
     private String email;
@@ -7,14 +7,16 @@ public class GetUserDetailsResponse {
     private String profilePhoto;
     private int tag;
     private String bio;
+    private boolean isAdmin;
 
-    public GetUserDetailsResponse(String email, String username, String name, String profilePhoto, int tag, String bio) {
+    public GetUserDetailsResponse(String email, String username, String name, String profilePhoto, int tag, String bio, boolean isAdmin) {
         this.email = email;
         this.username = username;
         this.name = name;
         this.profilePhoto = profilePhoto;
         this.tag = tag;
         this.bio = bio;
+        this.isAdmin = isAdmin;
     }
 
     public String getEmail() {
@@ -39,5 +41,9 @@ public class GetUserDetailsResponse {
 
     public String getBio() {
         return bio;
+    }
+
+    public boolean isAdmin() {
+        return isAdmin;
     }
 }
