+++ backend/src/main/java/com/bounswe2024group10/Tradeverse/model/User.java
@@ -128,6 +128,14 @@ public class User implements UserDetails {
         this.bio = bio;
     }
 
+    public boolean getIsAdmin() {
+        return isAdmin;
+    }
+
+    public void setIsAdmin(boolean isAdmin) {
+        this.isAdmin = isAdmin;
+    }
+
     // UserDetails interface methods
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
