+++ backend/src/main/java/com/bounswe2024group10/Tradeverse/service/LikeService.java
@@ -6,14 +6,7 @@ import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.bounswe2024group10.Tradeverse.dto.like.GetLikedPostsRequest;
-import com.bounswe2024group10.Tradeverse.dto.like.GetLikedPostsResponse;
-import com.bounswe2024group10.Tradeverse.dto.like.GetLikersRequest;
-import com.bounswe2024group10.Tradeverse.dto.like.GetLikersResponse;
-import com.bounswe2024group10.Tradeverse.dto.like.LikePostRequest;
-import com.bounswe2024group10.Tradeverse.dto.like.LikePostResponse;
-import com.bounswe2024group10.Tradeverse.dto.like.UnlikePostRequest;
-import com.bounswe2024group10.Tradeverse.dto.like.UnlikePostResponse;
+import com.bounswe2024group10.Tradeverse.dto.like.*;
 import com.bounswe2024group10.Tradeverse.model.Dislike;
 import com.bounswe2024group10.Tradeverse.model.Like;
 import com.bounswe2024group10.Tradeverse.model.Post;
